+++ frontend/src/containers/__tests__/Home-test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import Home from 'components/Home';
+import Home from '../Home';
 
 describe('home component', () => {
   it('matches snapshot', () => {
